make the Available Rides in Home for passenger static and not scrollable.
Also add a feature for the passenger to see the status of their requested rides.

The rules for new ride requests are as follows:
a) No request can be made for a ride before the current date and time
b) Passenger can post a maximum of 2 active ride requests.
c) If a ride request is not fulfilled by any driver till the time and date of the requested ride, the ride request is automatically deleted from both passenger and driver side

The driver must get a real time notification when a new ride request is submitted and the passenger is notified in real-time when his or her ride is accepted.

A new ride request can be either in sharing or in full-reserved type, in which case no other passenger can book a spot in the same ride. In case of sharing, each ride can have a maximum of 4 passengers.